requirements.txt
# requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
WTForms==3.1.2
werkzeug==3.0.4
python-dotenv==1.0.1